+++ fsm_test.go
@@ -266,7 +266,9 @@ func Test_FSM_TransitionWithHandlers(t *testing.T) {
 
 			if tt.err != nil {
 				assert.Error(t, err)
-				assert.EqualError(t, err, tt.err.Error())
+
+				// TODO: it is correctly?
+				assert.Contains(t, err.Error(), tt.err.Error())
 			} else {
 				assert.NoError(t, err)
 			}